refactor: Use `PlatformIDSet` instead of `[]PlatformID` for clarity
